+++ lib/open_fda_api/query_builder.rb
@@ -42,13 +42,9 @@ module OpenFdaApi
     def build_search_string(search)
       return "" if search.empty?
 
-      value = if search.count == 1 # AND case
-                # Turns [{a: b, c: d}] into "a:b+AND+c:d"
-                search.first.map { |k, v| "#{k}:#{v}" }.join("+AND+")
-              else # OR case
-                # Turns [{a: b}, {c: d}] into "a:b+c:d"
-                search.map { |h| "#{h.keys.first}:#{h.values.first}" }.join("+")
-              end
+      value = search.map do |and_args|
+        "(#{and_args.map { |k, v| "#{k}:#{v.gsub(" ", "+")}" }.join("+AND+")})"
+      end.join("+")
 
       "search=#{value}"
     end